Multichoice and SingelAnswer Features
Back button on multichoice and singleanswer activities will now ask the
user if they want to return to the main menu. This is done by having a
setter for the trailEnded boolean within the QuestionManager class and
setting it to true when you press yes on the dialog fragment.
Both intents now have an additional button that is initially invisible
which will let the user - once the correct answer is given in either
multichoice or single answer scan - continue to the next question.
Removed bar from top of all screens by making them extend
FragmentActivity instead of ActionBarActivity.
Buttons on multichoice activity will now be disabled once they have been
clicked once to prevent losing more points by pressing the same button
over again.
This is also true for the button giving the correct answer.
The QR-scanner button on the SingleAnswerActivity will now be
unclickable and green once the correct answer is given. It will also
show the 'Next Question' button below it.