Added QRFragment and changes to MultiChoiceActivity
New QRFragment added to the bottom-row QR-button. This fragment contains
two TextViews - a title and a hint and an ImageButton containing the
transparent QR-button resource. This has been scaled to three times the
normal size due to current resource file constraints. This ImageButton
will now start the QRActivity for result instead of the bottom-row
button.
The text contained in the TextViews has been added to the strings.xml
file.